Fixed erros and improved structure

+++ models/account.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey
+from sqlalchemy import Column, Integer, String, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 from db.database import Base 
 
@@ -9,7 +9,7 @@ class Account (Base):
 
     id = Column(Integer, primary_key=True, index=True)
     acc_no = Column(String(10), unique=True, nullable=False)
-    balance = Column(Float, default=0)
+    balance = Column((Numeric(10,2)), default=0)
     user_id = Column(Integer, ForeignKey("users.id"))
     owner = relationship("User", back_populates="accounts")
